feat : extract progress bar component
