+++ base-react/src/components/style.css
@@ -1,0 +1,8 @@
+.fix {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  background-color: white;
+  z-index: 100;
+}
